pytorch_afo_testkit-main/requirements.txt
deepspeed
numpy
pyyaml
sentencepiece
tokenizers
torch
transformers
pandas
xlsxwriter
prettytable
